addressing coderabbitai comments
