+++ app/client/src/ce/selectors/environmentSelectors.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import type { AppState } from "@appsmith/reducers";
 import type { EnvironmentType } from "@appsmith/configs/types";
+import { DEFAULT_ENV_ID } from "constants/EnvironmentContants";
 
 export const areEnvironmentsFetched = (state: AppState, workspaceId: string) =>
   true;
@@ -10,18 +11,19 @@ export const getEnvironmentsWithPermission = (
   state: AppState,
 ): Array<EnvironmentType> => [];
 
-export const getDefaultEnvironmentId = (state: AppState) => "unused_env";
+export const getDefaultEnvironmentId = (state: AppState) => DEFAULT_ENV_ID;
 
-export const getCurrentEnvironmentId = (state: AppState) => "unused_env";
+export const getCurrentEnvironmentId = (state: AppState) => DEFAULT_ENV_ID;
 
 export const getCurrentEnvironmentName = (state: AppState) => "";
 
-export const getCurrentEditingEnvironmentId = (state: AppState) => "unused_env";
+export const getCurrentEditingEnvironmentId = (state: AppState) =>
+  DEFAULT_ENV_ID;
 
 export const getCurrentEnvironmentDetails = (state: AppState) => ({
-  id: "unused_env",
+  id: DEFAULT_ENV_ID,
   name: "",
-  editingId: "unused_env",
+  editingId: DEFAULT_ENV_ID,
 });
 
 export const allowManageEnvironmentAccessForUser = (state: AppState) => false;
